cpu-vm: relax cpu pinning check a bit
It was discovered, that current check, while being correct
in theory works really bad in practice. Because QEMU sometimes
spawns additional temporary threads from an actual vCPU threads
which leads to the situation when we have more threads with pinning
applied than vCPU number. Let's relax our check a bit and check
for "greater than" and not for the equality.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>
Signed-off-by: hamistao <pedro.ribeiro@canonical.com>