tests/cgroup: take into account LXCFS support for swap accounting
Older LXCFS versions (4.0.x and 5.0.x branches) do not handle
swap accounting on cgroup2 properly. While we are improving cgroup2
tests for a newer LXD versions we don't want to make tests red on
the older ones. So, let's check if LXCFS declares support for
swap accounting and only in this case make any checks for it.

See also:
https://github.com/lxc/lxcfs/pull/663

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>